Create Startup class
In order to use dependency injection and configure HotChocolate, we use a functions startup class. It works like a regular startup class you'd write for asp.net and asp.net core, though it's autogenerated if not declared explicitly.

HotChocolate uses the IoC capabilities of asp.net core to configure and use services. They are injected into the resolvers so you can use them