Create the query type
The most basic type of every graphql schema is a type representing queries. This type exposes fields to query data from the API.

To get started, we creat a field called greeting, returning a greeting string. Next up we'll register the query type with HotChocolate and use it within our graphql endpoint